+++ R/gene_set_enrichment.R
@@ -198,14 +198,10 @@ gse_hypergeometric_list <- function(
 
 #### original implementations --------------------------------------------------
 
-# TODO Implement wrapper around the more traditional GSEA function
-
-#### simple implementation -----------------------------------------------------
-
 #' Bixverse implementation of the simple fgsea algorithm
 #'
 #' @description
-#' Rust-based version of the fgsea simple algorithm.
+#' Rust-based version of the traditional permutation-based GSEA algorithm.
 #'
 #' @param stats Named numeric vector. The gene level statistic.
 #' @param pathways List. A named list with each element containing the genes for
@@ -223,7 +219,7 @@ gse_hypergeometric_list <- function(
 #' @returns To be written.
 #'
 #' @export
-fgsea_simple = function(
+calc_gsea_traditional = function(
   stats,
   pathways,
   nperm = 2000L,
@@ -271,7 +267,94 @@ fgsea_simple = function(
 
   pathway_scores <- unlist(gsea_stat_res[, "es"])
 
-  permutations_res <- with(
+  permutations_res_traditional <- rs_calc_gsea_stat_traditional_batch(
+    stats = stats,
+    pathway_scores = pathway_scores,
+    pathway_sizes = as.integer(pathway_sizes),
+    iters = nperm,
+    seed = seed
+  ) %>%
+    data.table::setDT() %>%
+    .[, `:=`(
+      pathway_name = rownames(gsea_stat_res),
+      leading_edge = leading_edges
+    )]
+
+  return(permutations_res_traditional)
+}
+
+#### simple implementation -----------------------------------------------------
+
+#' Bixverse implementation of the simple fgsea algorithm
+#'
+#' @description
+#' Rust-based version of the fgsea simple algorithm.
+#'
+#' @param stats Named numeric vector. The gene level statistic.
+#' @param pathways List. A named list with each element containing the genes for
+#' this pathway.
+#' @param nperm Integer. Number of permutation tests. Defaults to `2000L`.
+#' @param gsea_params List. The GSEA parameters, see [bixverse::params_gsea()]
+#' wrapper function. This function generates a list containing:
+#' \itemize{
+#'  \item min_size - Integer. Minimum size for the gene sets.
+#'  \item max_size - Integer. Maximum size for the gene sets.
+#'  \item gsea_param - Float. The GSEA parameter. Defaults to `1.0`.
+#' }
+#' @param seed Random seed for reproducibility.
+#'
+#' @returns To be written.
+#'
+#' @export
+calc_fgsea_simple = function(
+  stats,
+  pathways,
+  nperm = 2000L,
+  gsea_params = params_gsea(),
+  seed = 123L
+) {
+  # Checks
+  checkmate::assertNumeric(stats, min.len = 3L, finite = TRUE)
+  checkmate::assertNames(names(stats))
+  checkmate::assertList(pathways, types = "character")
+  checkmate::assertNames(names(pathways))
+  assertGSEAParams(gsea_params)
+
+  c(stats, pathways_clean, pathway_sizes) %<-%
+    with(
+      gsea_params,
+      prep_stats_pathways(
+        stats = stats,
+        pathways = pathways,
+        min_size = min_size,
+        max_size = max_size
+      )
+    )
+
+  gsea_stat_res <- with(
+    gsea_params,
+    do.call(
+      rbind,
+      lapply(
+        pathways_clean,
+        rs_calc_gsea_stats,
+        stats = stats,
+        gsea_param = gsea_param,
+        return_leading_edge = TRUE
+      )
+    )
+  )
+
+  leading_edges <- mapply(
+    "[",
+    list(names(stats)),
+    gsea_stat_res[, "leading_edge"],
+    SIMPLIFY = FALSE
+  )
+
+  pathway_scores <- unlist(gsea_stat_res[, "es"])
+
+  permutations_res_simple <- with(
     gsea_params,
     rs_calc_gsea_stat_cumulative_batch(
       stats = stats,
@@ -281,40 +364,14 @@ fgsea_simple = function(
       seed = seed,
       gsea_param = gsea_param
     )
-  )
-
-  le_zero_mean <- permutations_res$le_zero_sum / permutations_res$le_zero
-  ge_zero_mean <- permutations_res$ge_zero_sum / permutations_res$ge_zero
-
-  nes <- data.table::fifelse(
-    (pathway_scores > 0 & ge_zero_mean != 0) |
-      (pathway_scores < 0 & le_zero_mean != 0),
-    pathway_scores /
-      data.table::fifelse(
-        pathway_scores > 0,
-        ge_zero_mean,
-        abs(le_zero_mean)
-      ),
-    NA
-  )
-
-  pvals <- pmin(
-    (1 + permutations_res$le_es) / (1 + permutations_res$le_zero),
-    (1 + permutations_res$ge_es) / (1 + permutations_res$ge_zero)
-  )
-
-  final_res <- data.table::as.data.table(
-    gsea_stat_res,
-    keep.rownames = "pathway"
   ) %>%
+    data.table::setDT() %>%
     .[, `:=`(
-      nes = nes,
-      pval = pvals,
-      leading_edge = leading_edges,
-      es = unlist(es)
+      pathway_name = rownames(gsea_stat_res),
+      leading_edge = leading_edges
     )]
 
-  return(final_res)
+  return(permutations_res_simple)
 }
 
 #### multi level implementation ------------------------------------------------